+++ .env.development
@@ -1,0 +1,5 @@
+// TODO 开发环境
+
+ENV = 'development'
+
+VUE_APP_BASE_URL = 'http://localhost:5000/'
