+++ .env.production
@@ -1,0 +1,4 @@
+// TODO 生产环境
+ENV = 'production'
+
+VUE_APP_BASE_URL = 'http://localhost:5000/'
