+++ .env.test
@@ -1,0 +1,2 @@
+// TODO 测试环境
+VUE_APP_BASE_URL = 'http://localhost:5000/'
