+++ src/assets/styles/index.scss
@@ -1,0 +1,1 @@
+@import './realize/normalize.scss'
