--- src/assets/styles/realize/element-variables.scss
@@ -1,7 +1,0 @@
-/* theme color */
-$--color-primary: teal;
-
-/* icon font path, required */
-$--font-path: '~element-ui/lib/theme-chalk/fonts';
-
-@import "~element-ui/packages/theme-chalk/src/index";
