+++ src/utils/request.js
@@ -1,0 +1,8 @@
+// import axios from 'axios'
+
+// const service = axios.create({
+//   baseURL: process.env.VUE_APP_BASE_URL,
+//   timeout: 12000,
+//   headers: {
+//   }
+// })
