+++ src/views/User/index.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="user">
+    User页面
+  </div>
+</template>
+
+<script type='text/ecmascript-6'>
+export default {
+  name: 'user',
+  data () {
+    return {}
+  }
+}
+</script>
+
+<style lang="scss" scoped rel="stylesheet/scss">
+</style>
