+++ vue.config.js
@@ -1,4 +1,16 @@
+'use strict'
+const path = require('path')
+
+const resolve = dir => path.join(__dirname, dir)
+
 module.exports = {
+  chainWebpack: (config) => {
+    config.resolve.alias
+      .set('@', resolve('src'))
+      .set('assets', resolve('src/assets'))
+      .set('components', resolve('src/components'))
+      .set('views', resolve('src/views'))
+  },
   devServer: {
     port: 8989
   }
